Move recurseAndRender from MobileContentPageRenderer into new class MobileContentViewRenderer.  This is now part of the renderer page context.
